pyproject.toml
[project]
name = "gemini-pdf-batch-analyzer"
version = "0.1.0"
description = "Batch PDF analysis with Google Gemini"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai",
    "pypdf",
    "pandas",
    "python-dotenv",
    "tqdm",
    "openpyxl",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "ruff",
]

[project.scripts]
pdf-analyzer = "gemini_pdf_analyzer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
